+++ mods/DeutschZ_Core/config.cpp
@@ -5,7 +5,7 @@ class CfgPatches
         units[] = {};
         weapons[] = {};
         requiredVersion = 0.1;
-        requiredAddons[] = {"DZ_Data"};
+        requiredAddons[] = {"DZ_Data", "DZ_Scripts"};
     };
 };
 
@@ -14,7 +14,7 @@ class CfgMods
     class DeutschZ_Core
     {
         dir = "DeutschZ_Core";
-        name = "DeutschZ_Core";
+        name = "DeutschZ Core";
         author = "Patrick Sluzalek / fck1701";
         type = "mod";
         dependencies[] = {"Game", "World", "Mission"};